Separate tls and non-tls server in test.
